requirements.txt
Flask==2.3.2
Flask-Dance==6.2.0
Flask-Limiter==3.3.1
Werkzeug==2.3.6
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0
pyOpenSSL==23.2.0
